// src/types/bindings/DashboardMetrics.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GstSummaryBreakdown } from "./GstSummaryBreakdown";

export type DashboardMetrics = { today_sales: number, mtd_sales: number, ytd_sales: number, comparative_growth_percent: number, pending_credit_notes_count: number, pending_debit_notes_count: number, cancelled_invoices_count: number, top_10_customers: Array<[string, number]>, top_10_suppliers: Array<[string, number]>, top_20_parts: Array<[string, number]>, gst_payable_summary: GstSummaryBreakdown, import_errors_count: number, recent_activity: Array<string>, };